Moved XR to Plugins Folder and enabled XR devices in Editor (will investigate other platforms).
