--- Packages/com.unity.inputsystem/Extras/XR.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5f4295264f636a84295850c592666e71
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
