--- Packages/com.unity.inputsystem/Extras/XR/Unity.InputSystem.XR.asmdef
@@ -1,10 +1,0 @@
-{
-    "name": "Unity.InputSystem.XR",
-    "references": [
-        "Unity.InputSystem"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": true
-}
