+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 75de1fe83d7195b4ab0fb4c635c83f11
-AssemblyDefinitionImporter:
+guid: 097288e93d94f44438e9ef4dc93f7346
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
