+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/Haptics/SimpleXRRumble.cs.meta
