+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/Haptics/SimpleXRRumbleCommand.cs.meta
