+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/XRTemplateBuilder.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Experimental.Input.LowLevel;
 using UnityEngine.Experimental.Input.Utilities;
-using UnityEngine.Experimental.Input.XR.Haptics;
+using UnityEngine.Experimental.Input.Plugins.XR.Haptics;
 using System.Text;
 
-namespace UnityEngine.Experimental.Input.XR
+namespace UnityEngine.Experimental.Input.Plugins.XR
 {
     [Serializable]
     class XRTemplateBuilder
