+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/XRTemplateBuilder.cs.meta
